requirements.txt
Django>=4.2.0,<5.0
django-import-export>=3.0.0
Pillow>=10.0.0
python-decouple>=3.8
django-crispy-forms>=2.0
crispy-bootstrap4>=2022.1
faker>=18.0.0
python-dateutil>=2.8.0
